app/services/db.py
# app/services/db.py
import os, pymysql
from dotenv import load_dotenv
load_dotenv()

def get_conn():
    conn = pymysql.connect(
        host=os.getenv("DB_HOST"),
        user=os.getenv("DB_USER"),
        password=os.getenv("DB_PASS"),
        database=os.getenv("DB_NAME"),
        port=int(os.getenv("DB_PORT", "3306")),
        cursorclass=pymysql.cursors.DictCursor,
        autocommit=True,
        charset="utf8mb4",
    )
    with conn.cursor() as cur:
        cur.execute("SET NAMES utf8mb4 COLLATE utf8mb4_unicode_ci;")
        cur.execute("SET collation_connection = 'utf8mb4_unicode_ci';")
    return conn

