requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
PyMySQL==1.1.1
bcrypt==4.1.3
watchfiles==0.23.0

